move httpx/fetch => internal/fetch
As part of the work to fully migrate to netx, let's also dismantle
the httpx directory which we don't need to be public.
