+++ internal/fetch/fetch_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/apex/log"
-	"github.com/ooni/probe-engine/httpx/fetch"
+	"github.com/ooni/probe-engine/internal/fetch"
 	"github.com/ooni/probe-engine/httpx/httplog"
 	"github.com/ooni/probe-engine/httpx/httptracex"
 )
